move from network 2019-07-01 to 2020-08-01
