+++ deploy/env-development.json
@@ -60,7 +60,7 @@
             "type": "Microsoft.Network/publicIPAddresses",
             "location": "[resourceGroup().location]",
             "condition": "[equals(parameters('ciCapacity'), 0)]",
-            "apiVersion": "2019-07-01"
+            "apiVersion": "2020-08-01"
         },
         {
             "properties": {
@@ -91,7 +91,7 @@
             "name": "dev-vnet",
             "type": "Microsoft.Network/virtualNetworks",
             "location": "[resourceGroup().location]",
-            "apiVersion": "2019-07-01"
+            "apiVersion": "2020-08-01"
         },
         {
             "properties": {
@@ -136,7 +136,7 @@
             "type": "Microsoft.Network/virtualNetworkGateways",
             "location": "[resourceGroup().location]",
             "condition": "[equals(parameters('ciCapacity'), 0)]",
-            "apiVersion": "2019-07-01",
+            "apiVersion": "2020-08-01",
             "dependsOn": [
                 "[resourceId('Microsoft.Network/publicIPAddresses', 'dev-vpn-pip')]",
                 "[resourceId('Microsoft.Network/virtualNetworks', 'dev-vnet')]"
