+++ deploy/rp-development.json
@@ -49,7 +49,7 @@
             "name": "rp-vnet",
             "type": "Microsoft.Network/virtualNetworks",
             "location": "[resourceGroup().location]",
-            "apiVersion": "2019-07-01"
+            "apiVersion": "2020-08-01"
         },
         {
             "properties": {
@@ -75,7 +75,7 @@
             "name": "rp-pe-vnet-001",
             "type": "Microsoft.Network/virtualNetworks",
             "location": "[resourceGroup().location]",
-            "apiVersion": "2019-07-01"
+            "apiVersion": "2020-08-01"
         },
         {
             "properties": {
@@ -89,7 +89,7 @@
             },
             "name": "rp-vnet/peering-rp-pe-vnet-001",
             "type": "Microsoft.Network/virtualNetworks/virtualNetworkPeerings",
-            "apiVersion": "2019-07-01",
+            "apiVersion": "2020-08-01",
             "dependsOn": [
                 "[resourceId('Microsoft.Network/virtualNetworks', 'rp-vnet')]",
                 "[resourceId('Microsoft.Network/virtualNetworks', 'rp-pe-vnet-001')]"
@@ -108,7 +108,7 @@
             },
             "name": "rp-pe-vnet-001/peering-rp-vnet",
             "type": "Microsoft.Network/virtualNetworks/virtualNetworkPeerings",
-            "apiVersion": "2019-07-01",
+            "apiVersion": "2020-08-01",
             "dependsOn": [
                 "[resourceId('Microsoft.Network/virtualNetworks', 'rp-pe-vnet-001')]",
                 "[resourceId('Microsoft.Network/virtualNetworks', 'rp-vnet')]"
