+++ pkg/util/azureclient/apiversions.go
@@ -20,7 +20,7 @@ var apiVersions = map[string]string{
 	"microsoft.insights":                      "2018-03-01",
 	"microsoft.keyvault":                      "2019-09-01",
 	"microsoft.managedidentity":               "2018-11-30",
-	"microsoft.network":                       "2019-07-01",
+	"microsoft.network":                       "2020-08-01",
 	"microsoft.network/dnszones":              "2018-05-01",
 	"microsoft.network/privatednszones":       "2018-09-01",
 	"microsoft.storage":                       "2019-04-01",
