+++ pkg/util/azureclient/mgmt/network/privateendpoints_addons.go
@@ -6,7 +6,7 @@ package network
 import (
 	"context"
 
-	mgmtnetwork "github.com/Azure/azure-sdk-for-go/services/network/mgmt/2019-07-01/network"
+	mgmtnetwork "github.com/Azure/azure-sdk-for-go/services/network/mgmt/2020-08-01/network"
 )
 
 // PrivateEndpointsClientAddons contains addons for PrivateEndpointsClient
