+++ pkg/util/azureclient/mgmt/network/virtualnetworkpeerings.go
@@ -4,7 +4,7 @@ package network
 // Licensed under the Apache License 2.0.
 
 import (
-	mgmtnetwork "github.com/Azure/azure-sdk-for-go/services/network/mgmt/2019-07-01/network"
+	mgmtnetwork "github.com/Azure/azure-sdk-for-go/services/network/mgmt/2020-08-01/network"
 	"github.com/Azure/go-autorest/autorest"
 
 	"github.com/Azure/ARO-RP/pkg/util/azureclient"
