+++ pkg/util/mocks/subnet/subnet.go
@@ -8,7 +8,7 @@ import (
 	context "context"
 	reflect "reflect"
 
-	network "github.com/Azure/azure-sdk-for-go/services/network/mgmt/2019-07-01/network"
+	network "github.com/Azure/azure-sdk-for-go/services/network/mgmt/2020-08-01/network"
 	gomock "github.com/golang/mock/gomock"
 )
